python_requirements_compatibility_check/requirements.txt
requirements-parser
requests